+++ Src/FluentAssertions/Specialized/AsyncFunctionAssertions.cs
@@ -55,7 +55,7 @@ public class AsyncFunctionAssertions<TTask, TAssertions> : DelegateAssertionsBas
 
             if (remainingTime >= TimeSpan.Zero)
             {
-                bool completesWithinTimeout = await CompletesWithinTimeoutAsync(task, remainingTime);
+                bool completesWithinTimeout = await CompletesWithinTimeoutAsync(task, remainingTime, _ => Task.CompletedTask);
 
                 assertionChain
                     .ForCondition(!completesWithinTimeout)
@@ -228,7 +228,7 @@ public class AsyncFunctionAssertions<TTask, TAssertions> : DelegateAssertionsBas
                 // Here we do not need to know whether the task completes (successfully) in timeout
                 // or does not complete. We are only interested in the exception which is thrown, not returned.
                 // So, we can ignore the result.
-                _ = await CompletesWithinTimeoutAsync(task, remainingTime);
+                _ = await CompletesWithinTimeoutAsync(task, remainingTime, cancelledTask => cancelledTask);
             }
 
             return null;
@@ -289,7 +289,7 @@ public class AsyncFunctionAssertions<TTask, TAssertions> : DelegateAssertionsBas
     /// <summary>
     ///     Monitors the specified task whether it completes withing the remaining time span.
     /// </summary>
-    private protected async Task<bool> CompletesWithinTimeoutAsync(Task target, TimeSpan remainingTime)
+    private protected async Task<bool> CompletesWithinTimeoutAsync(Task target, TimeSpan remainingTime, Func<Task, Task> onTaskCanceled)
     {
         using var delayCancellationTokenSource = new CancellationTokenSource();
 
@@ -310,8 +310,7 @@ public class AsyncFunctionAssertions<TTask, TAssertions> : DelegateAssertionsBas
 
         if (target.IsCanceled)
         {
-            // Rethrow the exception causing the task be canceled.
-            await target;
+            await onTaskCanceled(target);
         }
 
         // The monitored task is completed, we shall cancel the clock.