+++ Src/FluentAssertions/Specialized/NonGenericAsyncFunctionAssertions.cs
@@ -62,7 +62,7 @@ public class NonGenericAsyncFunctionAssertions : AsyncFunctionAssertions<Task, N
 
             if (assertionChain.Succeeded)
             {
-                bool completesWithinTimeout = await CompletesWithinTimeoutAsync(task, remainingTime);
+                bool completesWithinTimeout = await CompletesWithinTimeoutAsync(task, remainingTime, _ => Task.CompletedTask);
 
                 assertionChain
                     .ForCondition(completesWithinTimeout)